+++ examples/plotting.py
@@ -78,12 +78,31 @@ def confusion_matrix_example():
     return fig, ax
 
 
-if __name__ == "__main__":
+def spectral_distortion_index_example():
+    from torchmetrics.image.d_lambda import SpectralDistortionIndex
 
+    p = lambda: torch.rand([16, 3, 16, 16])
+    t = lambda: torch.rand([16, 3, 16, 16])
+
+    # plot single value
+    metric = SpectralDistortionIndex()
+    metric.update(p(), t())
+    fig, ax = metric.plot()
+
+    # plot multiple values
+    metric = SpectralDistortionIndex()
+    vals = [metric(p(), t()) for _ in range(10)]
+    fig, ax = metric.plot(vals)
+
+    return fig, ax
+
+
+if __name__ == "__main__":
     metrics_func = {
         "accuracy": accuracy_example,
         "mean_squared_error": mean_squared_error_example,
         "confusion_matrix": confusion_matrix_example,
+        "spectral_distortion_index": spectral_distortion_index_example,
     }
 
     parser = argparse.ArgumentParser(description="Example script for plotting metrics.")